+++ license_tools/__main__.py
@@ -6,7 +6,7 @@ from __future__ import annotations
 
 import argparse
 
-from license_tools import scancode_tools
+from license_tools import retrieval
 
 
 def main() -> None:
@@ -16,38 +16,78 @@ def main() -> None:
 
     source_group = parser.add_argument_group("Artifact source")
     source_group = source_group.add_mutually_exclusive_group(required=False)
-    source_group.add_argument("--directory", action="store", type=str, help="Directory to work on.")
-    source_group.add_argument("--file", action="store", type=str, help="File to work on.")
-    source_group.add_argument("--archive", action="store", type=str, help="Archive file to work on.")
-    source_group.add_argument("--package", action="store", type=str, help="Package specification to use.")
-    source_group.add_argument("--url", action="store", type=str, help="Download URL to use.")
+    source_group.add_argument(
+        "--directory", action="store", type=str, help="Directory to work on."
+    )
+    source_group.add_argument(
+        "--file", action="store", type=str, help="File to work on."
+    )
+    source_group.add_argument(
+        "--archive", action="store", type=str, help="Archive file to work on."
+    )
+    source_group.add_argument(
+        "--package", action="store", type=str, help="Package specification to use."
+    )
+    source_group.add_argument(
+        "--url", action="store", type=str, help="Download URL to use."
+    )
 
     parser.add_argument(
-        "--index-url", action="store", type=str, required=False, default="", help="PyPI index URL to use."
+        "--index-url",
+        action="store",
+        type=str,
+        required=False,
+        default="",
+        help="PyPI index URL to use.",
     )
     parser.add_argument(
-        "--jobs", action="store", type=int, required=False, default=4, help="Parallel jobs to use."
+        "--jobs",
+        action="store",
+        type=int,
+        required=False,
+        default=4,
+        help="Parallel jobs to use.",
     )
 
     parser.add_argument(
-        "--retrieve-copyrights", action="store_true", required=False, default=False, help="Retrieve copyrights."
+        "--retrieve-copyrights",
+        action="store_true",
+        required=False,
+        default=False,
+        help="Retrieve copyrights.",
     )
     parser.add_argument(
-        "--retrieve-emails", action="store_true", required=False, default=False, help="Retrieve e-mails."
+        "--retrieve-emails",
+        action="store_true",
+        required=False,
+        default=False,
+        help="Retrieve e-mails.",
     )
     parser.add_argument(
-        "--retrieve-file-info", action="store_true", required=False, default=False, help="Retrieve file information."
+        "--retrieve-file-info",
+        action="store_true",
+        required=False,
+        default=False,
+        help="Retrieve file information.",
     )
     parser.add_argument(
-        "--retrieve-urls", action="store_true", required=False, default=False, help="Retrieve URLs."
+        "--retrieve-urls",
+        action="store_true",
+        required=False,
+        default=False,
+        help="Retrieve URLs.",
     )
     parser.add_argument(
-        "--retrieve-ldd-data", action="store_true", required=False, default=False, help="Retrieve shared object linking data."
+        "--retrieve-ldd-data",
+        action="store_true",
+        required=False,
+        default=False,
+        help="Retrieve shared object linking data.",
     )
 
     arguments = parser.parse_args()
 
-    scancode_tools.run(
+    retrieval.run(
         directory=arguments.directory,
         file_path=arguments.file,
         archive_path=arguments.archive,